main screen data gets updated now;

+++ source_apps/locationTracker/app/src/main/java/com/example/locationtracker/model/SyncInfo.kt
@@ -3,5 +3,5 @@ data class SyncInfo(
     var lastSync: String,
     val eventsNotSynced: Int,
     val oldestEventTimeNotSynced: String,
-    val totalEvents: Long
+    val totalEvents: Int
 )
